pyproject.toml
[project]
name = "whisper-bench"
version = "0.1.0"
description = "Comprehensive Whisper model benchmarking and comparison tool with performance optimization support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlx-whisper",
    "openai-whisper",
    "faster-whisper",
    "lightning-whisper-mlx",
    "librosa",
    "numpy",
    "soundfile>=0.13.1",
]

[project.optional-dependencies]
performance = [
    "torch>=2.0.0",  # For GPU detection and optimization
]

all = [
    "whisper-bench[performance]",
]

[project.scripts]
whisper-bench = "main:main"
